gripql/python: adding some unit tests
